New methods should not be in `hal-unproven`.
