+++ src/gpio/hal.rs
@@ -19,11 +19,14 @@
 // DEALINGS IN THE SOFTWARE.
 
 use core::convert::Infallible;
+use std::time::Duration;
 
 use embedded_hal::digital::OutputPin as OutputPinHal;
 use embedded_hal::pwm::PwmPin;
 
 use super::{Error, IoPin, OutputPin};
+
+const NANOS_PER_SEC: f64 = 1_000_000_000.0;
 
 impl OutputPinHal for OutputPin {
     type Error = Infallible;
@@ -81,6 +84,69 @@ impl embedded_hal_0::digital::v2::OutputPin for IoPin {
     }
 }
 
+impl embedded_hal::pwm::Pwm for OutputPin {
+    type Duty = f64;
+    type Channel = ();
+    type Time = Duration;
+    type Error = Error;
+
+    /// Disables a PWM `channel`
+    fn try_disable(&mut self, _channel: Self::Channel) -> Result<(), Self::Error> {
+        self.clear_pwm()
+    }
+
+    /// Enables a PWM `channel`
+    fn try_enable(&mut self, _channel: Self::Channel) -> Result<(), Self::Error> {
+        self.set_pwm_frequency(self.frequency, self.duty_cycle)
+    }
+
+    /// Returns the current PWM period
+    fn try_get_period(&self) -> Result<Self::Time, Self::Error> {
+        Ok(Duration::from_nanos(if self.frequency == 0.0 {
+            0
+        } else {
+            ((1.0 / self.frequency) * NANOS_PER_SEC) as u64
+        }))
+    }
+
+    /// Returns the current duty cycle
+    fn try_get_duty(&self, _channel: Self::Channel) -> Result<Self::Duty, Self::Error> {
+        Ok(self.duty_cycle)
+    }
+
+    /// Returns the maximum duty cycle value
+    fn try_get_max_duty(&self) -> Result<Self::Duty, Self::Error> {
+        Ok(1.0)
+    }
+
+    /// Sets a new duty cycle
+    fn try_set_duty(&mut self, _channel: Self::Channel, duty: Self::Duty) -> Result<(), Self::Error> {
+        self.duty_cycle = duty.max(0.0).min(1.0);
+
+        if self.soft_pwm.is_some() {
+            self.set_pwm_frequency(self.frequency, self.duty_cycle)?;
+        }
+
+        Ok(())
+    }
+
+    /// Sets a new PWM period
+    fn try_set_period<P>(&mut self, period: P) -> Result<(), Self::Error>
+    where
+        P: Into<Self::Time>,
+    {
+        let period = period.into();
+        self.frequency =
+            1.0 / (period.as_secs() as f64 + (f64::from(period.subsec_nanos()) / NANOS_PER_SEC));
+
+        if self.soft_pwm.is_some() {
+            self.set_pwm_frequency(self.frequency, self.duty_cycle)?;
+        }
+
+        Ok(())
+    }
+}
+
 impl PwmPin for OutputPin {
     type Duty = f64;
     type Error = Error;
@@ -133,6 +199,69 @@ impl embedded_hal_0::PwmPin for OutputPin {
 
     fn set_duty(&mut self, duty: Self::Duty) {
         let _ = self.try_set_duty(duty);
+    }
+}
+
+impl embedded_hal::pwm::Pwm for IoPin {
+    type Duty = f64;
+    type Channel = ();
+    type Time = Duration;
+    type Error = Error;
+
+    /// Disables a PWM `channel`
+    fn try_disable(&mut self, _channel: Self::Channel) -> Result<(), Self::Error> {
+        self.clear_pwm()
+    }
+
+    /// Enables a PWM `channel`
+    fn try_enable(&mut self, _channel: Self::Channel) -> Result<(), Self::Error> {
+        self.set_pwm_frequency(self.frequency, self.duty_cycle)
+    }
+
+    /// Returns the current PWM period
+    fn try_get_period(&self) -> Result<Self::Time, Self::Error> {
+        Ok(Duration::from_nanos(if self.frequency == 0.0 {
+            0
+        } else {
+            ((1.0 / self.frequency) * NANOS_PER_SEC) as u64
+        }))
+    }
+
+    /// Returns the current duty cycle
+    fn try_get_duty(&self, _channel: Self::Channel) -> Result<Self::Duty, Self::Error> {
+        Ok(self.duty_cycle)
+    }
+
+    /// Returns the maximum duty cycle value
+    fn try_get_max_duty(&self) -> Result<Self::Duty, Self::Error> {
+        Ok(1.0)
+    }
+
+    /// Sets a new duty cycle
+    fn try_set_duty(&mut self, _channel: Self::Channel, duty: Self::Duty) -> Result<(), Self::Error> {
+        self.duty_cycle = duty.max(0.0).min(1.0);
+
+        if self.soft_pwm.is_some() {
+            self.set_pwm_frequency(self.frequency, self.duty_cycle)?;
+        }
+
+        Ok(())
+    }
+
+    /// Sets a new PWM period
+    fn try_set_period<P>(&mut self, period: P) -> Result<(), Self::Error>
+    where
+        P: Into<Self::Time>,
+    {
+        let period = period.into();
+        self.frequency =
+            1.0 / (period.as_secs() as f64 + (f64::from(period.subsec_nanos()) / NANOS_PER_SEC));
+
+        if self.soft_pwm.is_some() {
+            self.set_pwm_frequency(self.frequency, self.duty_cycle)?;
+        }
+
+        Ok(())
     }
 }
