+++ src/gpio/hal_unproven.rs
@@ -18,7 +18,6 @@
 // FROM, OUT OF OR IN CONNECTION WITH THE SOFTWARE OR THE USE OR OTHER
 // DEALINGS IN THE SOFTWARE.
 
-use std::time::Duration;
 use core::convert::Infallible;
 
 use embedded_hal::digital::{
@@ -26,11 +25,8 @@ use embedded_hal::digital::{
     StatefulOutputPin as StatefulOutputPinHal,
     ToggleableOutputPin as ToggleableOutputPinHal,
 };
-use embedded_hal::pwm::Pwm;
 
 use super::{InputPin, IoPin, Level, OutputPin, Pin, Error};
-
-const NANOS_PER_SEC: f64 = 1_000_000_000.0;
 
 impl InputPinHal for Pin {
     type Error = Infallible;
@@ -204,69 +200,6 @@ impl embedded_hal_0::digital::v2::ToggleableOutputPin for OutputPin {
     }
 }
 
-impl Pwm for OutputPin {
-    type Duty = f64;
-    type Channel = ();
-    type Time = Duration;
-    type Error = Error;
-
-    /// Disables a PWM `channel`
-    fn try_disable(&mut self, _channel: Self::Channel) -> Result<(), Self::Error> {
-        self.clear_pwm()
-    }
-
-    /// Enables a PWM `channel`
-    fn try_enable(&mut self, _channel: Self::Channel) -> Result<(), Self::Error> {
-        self.set_pwm_frequency(self.frequency, self.duty_cycle)
-    }
-
-    /// Returns the current PWM period
-    fn try_get_period(&self) -> Result<Self::Time, Self::Error> {
-        Ok(Duration::from_nanos(if self.frequency == 0.0 {
-            0
-        } else {
-            ((1.0 / self.frequency) * NANOS_PER_SEC) as u64
-        }))
-    }
-
-    /// Returns the current duty cycle
-    fn try_get_duty(&self, _channel: Self::Channel) -> Result<Self::Duty, Self::Error> {
-        Ok(self.duty_cycle)
-    }
-
-    /// Returns the maximum duty cycle value
-    fn try_get_max_duty(&self) -> Result<Self::Duty, Self::Error> {
-        Ok(1.0)
-    }
-
-    /// Sets a new duty cycle
-    fn try_set_duty(&mut self, _channel: Self::Channel, duty: Self::Duty) -> Result<(), Self::Error> {
-        self.duty_cycle = duty.max(0.0).min(1.0);
-
-        if self.soft_pwm.is_some() {
-            self.set_pwm_frequency(self.frequency, self.duty_cycle)?;
-        }
-
-        Ok(())
-    }
-
-    /// Sets a new PWM period
-    fn try_set_period<P>(&mut self, period: P) -> Result<(), Self::Error>
-    where
-        P: Into<Self::Time>,
-    {
-        let period = period.into();
-        self.frequency =
-            1.0 / (period.as_secs() as f64 + (f64::from(period.subsec_nanos()) / NANOS_PER_SEC));
-
-        if self.soft_pwm.is_some() {
-            self.set_pwm_frequency(self.frequency, self.duty_cycle)?;
-        }
-
-        Ok(())
-    }
-}
-
 impl embedded_hal_0::Pwm for OutputPin {
     type Duty = f64;
     type Channel = ();
@@ -308,69 +241,6 @@ impl embedded_hal_0::Pwm for OutputPin {
         P: Into<Self::Time>,
     {
         let _ = self.try_set_period(period);
-    }
-}
-
-impl Pwm for IoPin {
-    type Duty = f64;
-    type Channel = ();
-    type Time = Duration;
-    type Error = Error;
-
-    /// Disables a PWM `channel`
-    fn try_disable(&mut self, _channel: Self::Channel) -> Result<(), Self::Error> {
-        self.clear_pwm()
-    }
-
-    /// Enables a PWM `channel`
-    fn try_enable(&mut self, _channel: Self::Channel) -> Result<(), Self::Error> {
-        self.set_pwm_frequency(self.frequency, self.duty_cycle)
-    }
-
-    /// Returns the current PWM period
-    fn try_get_period(&self) -> Result<Self::Time, Self::Error> {
-        Ok(Duration::from_nanos(if self.frequency == 0.0 {
-            0
-        } else {
-            ((1.0 / self.frequency) * NANOS_PER_SEC) as u64
-        }))
-    }
-
-    /// Returns the current duty cycle
-    fn try_get_duty(&self, _channel: Self::Channel) -> Result<Self::Duty, Self::Error> {
-        Ok(self.duty_cycle)
-    }
-
-    /// Returns the maximum duty cycle value
-    fn try_get_max_duty(&self) -> Result<Self::Duty, Self::Error> {
-        Ok(1.0)
-    }
-
-    /// Sets a new duty cycle
-    fn try_set_duty(&mut self, _channel: Self::Channel, duty: Self::Duty) -> Result<(), Self::Error> {
-        self.duty_cycle = duty.max(0.0).min(1.0);
-
-        if self.soft_pwm.is_some() {
-            self.set_pwm_frequency(self.frequency, self.duty_cycle)?;
-        }
-
-        Ok(())
-    }
-
-    /// Sets a new PWM period
-    fn try_set_period<P>(&mut self, period: P) -> Result<(), Self::Error>
-    where
-        P: Into<Self::Time>,
-    {
-        let period = period.into();
-        self.frequency =
-            1.0 / (period.as_secs() as f64 + (f64::from(period.subsec_nanos()) / NANOS_PER_SEC));
-
-        if self.soft_pwm.is_some() {
-            self.set_pwm_frequency(self.frequency, self.duty_cycle)?;
-        }
-
-        Ok(())
     }
 }
